analyze_slack_messageをpluginsに移動
